requirements_raw.txt
pandas==2.2.3
numpy==2.0.2
scikit-learn
ipykernel==6.29.5
tensorflow[and-cuda]==2.18.0; sys_platform == "linux"
tensorflow-cpu==2.18.0; sys_platform == "win32"
matplotlib==3.10.0
telemetry-parser==0.3.0
av==14.0.1
flask==3.1.0
python-dotenv==1.0.1
